Models: Consolidate to ONE ElementDefinition class
