+++ lib/logical/StructureDefinition.js
@@ -1,10 +1,10 @@
 const cloneDeep = require('lodash/cloneDeep');
-const TrackingElementDefinition = require('./TrackingElementDefinition');
+const ElementDefinition = require('./ElementDefinition');
 
 class StructureDefinition {
   constructor() {
     // Every structure definition needs a root element
-    const root = new TrackingElementDefinition('');
+    const root = new ElementDefinition('');
     root.structDef = this;
     root.mustSupport = false;
     root.isModifier = false;
@@ -133,7 +133,7 @@ class StructureDefinition {
   toJSON() {
     const j = { resourceType: 'StructureDefinition' };
     // First handle properties that are just straight translations to JSON
-    for (const prop of StructureDefinition._PROPS) {
+    for (const prop of PROPS) {
       if (this[prop] !== undefined) {
         j[prop] = cloneDeep(this[prop]);
       }
@@ -147,7 +147,7 @@ class StructureDefinition {
   static fromJSON(json) {
     const sd = new StructureDefinition();
     // First handle properties that are just straight translations from JSON
-    for (const prop of StructureDefinition._PROPS) {
+    for (const prop of PROPS) {
       if (json[prop] !== undefined) {
         sd[prop] = cloneDeep(json[prop]);
       }
@@ -155,14 +155,14 @@ class StructureDefinition {
     // Now handle the snapshots and (for now) just throw away the differential
     if (json.snapshot && json.snapshot.element) {
       for (const el of json.snapshot.element) {
-        sd._elements.push(TrackingElementDefinition.fromJSON(el));
+        sd._elements.push(ElementDefinition.fromJSON(el));
       }
     }
     return sd;
   }
 }
 
-StructureDefinition._PROPS = [ 'id', 'meta', 'implicitRules', 'language', 'text', 'contained', 'extension',
+const PROPS = [ 'id', 'meta', 'implicitRules', 'language', 'text', 'contained', 'extension',
   'modifierExtension', 'url', 'identifier', 'version', 'name', 'title', 'status', 'experimental', 'date',
   'publisher', 'contact', 'description', 'useContext', 'jurisdiction', 'purpose', 'copyright', 'keyword',
   'fhirVersion', 'mapping', 'kind', 'abstract', 'contextType', 'context', 'contextInvariant', 'type',
